Add undocumented synonym
I got annoyed recently that you assign editors, but you have to add reviewers. In my head, assigning and adding are the same thing. 

I want a way where you can also assign a reviewer, without having to type another command. This adds that possibility, but it doesn't enforce it for others by documenting or showing it as an option - which might be a good idea. Instead, it just will fall back on adding the user if someone mistakenly uses the word 'assign'.